refactor: 게임 종료 모달 컴포넌트화
